SnapshotTransactionManager -> SnapshotTransactionManagerImpl
